requirements.txt
# Core dependencies
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
httpx>=0.27.0
aiohttp>=3.9.0
aiofiles>=23.2.1

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.1

# Data & metrics
numpy>=1.26.0
pandas>=2.2.0

# Plotting
matplotlib>=3.8.0
seaborn>=0.13.0

# Async utilities
anyio>=4.4.0
tenacity>=8.3.0

# Logging
structlog>=24.2.0
python-json-logger>=2.0.7

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-timeout>=2.3.0
httpx>=0.27.0

# Optional: HuggingFace transformers backend
# transformers>=4.40.0
# torch>=2.3.0
# accelerate>=0.30.0

# Optional: vLLM backend (GPU only)
# vllm>=0.4.0

# Optional: Ray Serve
# ray[serve]>=2.20.0

# Process management
psutil>=5.9.0
